+++ acmed/src/main.rs
@@ -1,5 +1,4 @@
 use clap::{App, Arg};
-use env_logger::Builder;
 use log::{error, LevelFilter};
 
 mod acmed;
@@ -7,8 +6,10 @@ mod config;
 mod encoding;
 mod errors;
 mod hooks;
+mod logs;
 mod storage;
 
+pub const APP_NAME: &str = "acmed";
 pub const DEFAULT_CONFIG_FILE: &str = "/etc/acmed/acmed.toml";
 pub const DEFAULT_ACCOUNTS_DIR: &str = "/etc/acmed/accounts";
 pub const DEFAULT_CERT_DIR: &str = "/etc/acmed/certs";
@@ -20,6 +21,8 @@ pub const DEFAULT_POOL_TIME: u64 = 5000;
 pub const DEFAULT_CERT_FILE_MODE: u32 = 0o644;
 pub const DEFAULT_PK_FILE_MODE: u32 = 0o600;
 pub const DEFAULT_KP_REUSE: bool = false;
+pub const DEFAULT_LOG_SYSTEM: logs::LogSystem = logs::LogSystem::SysLog;
+pub const DEFAULT_LOG_LEVEL: LevelFilter = LevelFilter::Warn;
 
 fn main() {
     let matches = App::new("acmed")
@@ -40,30 +43,31 @@ fn main() {
                 .value_name("LEVEL")
                 .possible_values(&["error", "warn", "info", "debug", "trace"]),
         )
+        .arg(
+            Arg::with_name("to-syslog")
+                .long("log-syslog")
+                .help("Send log messages via syslog.")
+                .conflicts_with("to-stderr"),
+        )
+        .arg(
+            Arg::with_name("to-stderr")
+                .long("log-stderr")
+                .help("Print log messages to the standard error output.")
+                .conflicts_with("log-syslog"),
+        )
         .get_matches();
 
-    let mut builder = Builder::from_env("ACMED_LOG_LEVEL");
-    if let Some(v) = matches.value_of("log-level") {
-        match v {
-            "error" => {
-                builder.filter_level(LevelFilter::Error);
-            }
-            "warn" => {
-                builder.filter_level(LevelFilter::Warn);
-            }
-            "info" => {
-                builder.filter_level(LevelFilter::Info);
-            }
-            "debug" => {
-                builder.filter_level(LevelFilter::Debug);
-            }
-            "trace" => {
-                builder.filter_level(LevelFilter::Trace);
-            }
-            _ => {}
+    match logs::set_log_system(
+        matches.value_of("log-level"),
+        matches.is_present("log-syslog"),
+        matches.is_present("to-stderr"),
+    ) {
+        Ok(_) => {}
+        Err(e) => {
+            eprintln!("Error: {}", e);
+            std::process::exit(1);
         }
     };
-    builder.init();
 
     let config_file = matches.value_of("config").unwrap_or(DEFAULT_CONFIG_FILE);
     let mut srv = match acmed::Acmed::new(&config_file) {